requirements.txt
numpy>=1.24.0
aiohttp>=3.8.0
tqdm>=4.65.0
transformers>=4.30.0
cryptography>=40.0.0
psutil>=5.9.0
matplotlib>=3.7.0
pillow>=9.5.0
pyinstaller>=6.0.0 